generator/requirements.txt
kafka-python==2.0.2
numpy>=1.26.0
scipy>=1.11.0